# src/swagger/swagger.yaml
openapi: 3.0.4
info:
  title: Iarnród Éireann (Irish Rail) REST API v1
  description: >
    An easy-to-use REST API built on top of Irish Rail's SOAP API, designed for easy integration and JSON readability.
    Irish Rail's original SOAP API can be found at [api.irishrail.ie](https://api.irishrail.ie/realtime).
  version: 1.0.0
  contact:
    name: "TheWiredDiabetic"
    url: "https://thediabetic.dev"
    email: "taylor.thedev@proton.me"

servers:
  - url: http://localhost:3000
    description: Local server
  - url: https://dev.thediabetic.dev
    description: Development sandbox server
  - url: https://api.thediabetic.dev
    description: Production server

components:
  schemas:
    "Enums.LocationType":
      type: string
      description: |
        The type of location, usually seen on train movement JSONs.
        - **O**: Origin
        - **S**: Stop
        - **T**: Timing Point (Non-Stopping Location)
        - **D**: Destination
      enum: [O, S, T, D]
      example: S

    "Enums.StopType":
      type: string
      description: |
        The type of stop, usually seen on train movements for scheduled stops.
        - **C**: Current
        - **N**: Next
      enum: [C, N]
      example: C

    "Enums.TrainType":
      type: string
      enum: [A, D, S, M]
      description: |
        The type of train or route.
        - **A**: All
        - **D**: DART
        - **S**: Suburban (Commuter, etc.)
        - **M**: Mainline
      example: D

    "Enums.TrainStatus":
      type: string
      enum: [N, R]
      description: |
        The status of a current train.
        - **N**: Not Yet Running
        - **R**: Running
      example: R

    HaconTrain:
      type: object
      properties:
        code:
          type: string
          description: The unique 4-character code generated by Irish Rail.
          example: D123
        origin:
          type: object
          description: The origin station of the train.
          properties:
            name:
              type: string
              example: "Dublin Connolly"
            code:
              type: string
              example: "CNLLY"
            time:
              type: string
              format: date-time
              example: "2025-10-28T14:30:00Z"
            status:
              $ref: "#/components/schemas/Enums.TrainStatus"
            difference:
              type: number
              example: 0
            latitude:
              type: number
              format: float
              example: 51.8491
            longitude:
              type: number
              format: float
              example: -8.29956
        destination:
          type: object
          description: The destination station of the train.
          properties:
            name:
              type: string
              example: "Belfast"
            code:
              type: string
              example: "BFSTC"
            time:
              type: string
              format: date-time
              example: "2025-10-28T16:30:00Z"

    Train:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/Enums.TrainStatus"
        code:
          type: string
          description: The unique code generated by Irish Rail.
          example: E401
        date:
          type: string
          format: date
          description: The date of the service.
          example: 2025-10-28
        public_message:
          type: string
          description: >
            The public message of the train reporting its code, route, delay, and current status.
          example: "E401 Dublin Heuston to Cork - Departed Kildare (5 mins late)"
        direction:
          type: string
          description: The direction the service is operating (e.g., "To Cork").
          example: "To Cork"
        latitude:
          type: number
          format: float
          example: 53.349805
        longitude:
          type: number
          format: float
          example: -6.26031

    TrainMovement:
      type: object
      properties:
        code:
          type: string
          example: E936
          description: "The 4-character code generated by Irish Rail."
        date:
          type: string
          format: date
          example: "2025-10-28"
        location_code:
          type: string
          example: "BRAY"
        location_full_name:
          type: string
          example: "Bray"
        location_order:
          type: number
          example: 1
        location_type:
          $ref: "#/components/schemas/Enums.LocationType"
        train_origin:
          type: string
          example: "Bray"
        train_destination:
          type: string
          example: "Howth"
        scheduled_arrival:
          type: string
          example: "00:00:00"
        scheduled_departure:
          type: string
          example: "16:55:00"
        expected_arrival:
          type: string
          example: "00:00:00"
        expected_departure:
          type: string
          example: "00:00:00"
        arrival:
          type: string
          example: "17:10:00"
        departure:
          type: string
          example: "17:12:54"
        auto_arrival:
          type: string
          description: "Unknown use, blank for most services."
        auto_depart:
          type: string
          description: "Unknown use, blank for most services."
        stop_type:
          $ref: "#/components/schemas/Enums.StopType"

    StationObj:
      type: object
      properties:
        name:
          type: string
          example: "Clondalkin"
        code:
          type: string
          example: "CLDKN"
        alias:
          type: string
          example: "Fonthill ( Clondalkin )"
        latitude:
          type: number
          format: float
          example: 53.3334
        longitude:
          type: number
          format: float
          example: -6.40628

    StationTimetable:
      type: object
      properties:
        station_full_name:
          type: string
          example: "Dublin Connolly"
        station_code:
          type: string
          example: "CNLLY"
        destination:
          type: string
          example: "Howth"
        destination_time:
          type: string
          example: "16:10"
        status:
          type: string
          example: "En Route"
        last_location:
          type: string
          example: "Departed Grand Canal Dock"
        location_type:
          $ref: "#/components/schemas/Enums.LocationType"
        due_in:
          type: number
          example: 9
        late:
          type: number
          example: 10
        exp_arrival:
          type: string
          example: "15:53"
        exp_depart:
          type: string
          example: "15:54"
        sch_arrival:
          type: string
          example: "15:42"
        sch_departure:
          type: string
          example: "15:43"
        direction:
          type: string
          example: "Northbound"
        query_time:
          type: string
          format: date-time
          example: "2025-10-28T15:45:59Z"
        train_code:
          type: string
          example: "E930"
        train_date:
          type: string
          format: date
          example: "2025-10-28"
        train_type:
          type: string
          example: "DART"

    StationType:
      type: string
      enum: [D, S, A]
      description: |
        Service Type: 
        - **D** = DART
        - **S** = Suburban (Commuter & InterCity)
        - **A** = All (Unfiltered, default)
      example: D

    RouteError:
      type: object
      properties:
        success:
          type: boolean
          default: false
        errorCode:
          type: number
          example: 404
        errorMessage:
          type: string
          example: "No station found matching code 'BFSTA'"

tags:
  - name: trains
    description: Endpoints relating to trains
  - name: stations
    description: Endpoints relating to stations

paths:
  /trains:
    get:
      tags: [trains]
      summary: Get all running trains
      parameters:
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/Enums.TrainType"
          description: Filter trains by type
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  trains:
                    type: array
                    items:
                      $ref: "#/components/schemas/Train"

  /trains/{code}:
    get:
      tags: [trains]
      summary: Get a specific train by its code
      parameters:
        - name: code
          in: path
          required: true
          description: The train's 4-character code
          schema:
            type: string
            example: A114
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  train:
                    $ref: "#/components/schemas/Train"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteError"

  /trains/{code}/movements:
    get:
      tags: [trains]
      summary: Get movements for a specific train by its code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            example: A114
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  movements:
                    type: array
                    items:
                      $ref: "#/components/schemas/TrainMovement"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteError"

  /hacon-trains:
    get:
      tags: [trains]
      summary: Get all (Hacon) trains (used by the Irish Rail app)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HaconTrain"

  /stations:
    get:
      tags: [stations]
      summary: Get all stations operated by Irish Rail (incl. NI stations)
      parameters:
        - name: type
          in: query
          description: Filter stations by type
          schema:
            $ref: "#/components/schemas/StationType"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  stations:
                    type: array
                    items:
                      $ref: "#/components/schemas/StationObj"

  /stations/{code}:
    get:
      tags: [stations]
      summary: Get station by code
      parameters:
        - name: code
          in: path
          required: true
          description: The 5-character station code (e.g., CNLLY)
          schema:
            type: string
            example: CLDKN
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  station:
                    $ref: "#/components/schemas/StationObj"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteError"

  /stations/{code}/timetable:
    get:
      tags: [stations]
      summary: Get station timetable (inbound/outbound services)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            example: CNLLY
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  timetable:
                    type: array
                    items:
                      $ref: "#/components/schemas/StationTimetable"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteError"